convert Home.jsx to typescript
